Adds runtime logger class
