+++ core/impl-base/src/test/java/org/jboss/arquillian/core/impl/ManagerProcessingTestCase.java
@@ -32,7 +32,7 @@ import org.junit.Test;
 public class ManagerProcessingTestCase
 {
    static {
-      ManagerImpl.DEBUG = true;
+      RuntimeLogger.DEBUG = true;
    }
    
    @Test
